davinci_allocation/requirements.txt
Flask==2.0.1
pandas==1.3.3
openpyxl==3.0.9
python-dotenv==0.19.1
requests==2.26.0
Flask-WTF==1.0.0
email-validator==1.1.3
pytest==6.2.5 